# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Nextstep Dashboard API
  version: 1.2.0
  description: >-
    Contract for the multi-tenant SML dashboard, production administration,
    distributed report execution, LINE delivery, and LIFF viewer. Mutation
    requests authenticated by cookie also require X-CSRF-Token.
servers:
  - url: https://dashboard.nextstep-soft.com
...
paths:
...
  /api/v1/admin/tenants/{tenantId}/recipients/{recipientId}/agent-token:
    get:
      tags: [Admin configuration]
      operationId: getRecipientAgentToken
      description: Whether the assistant is enabled and the state of this recipient's token. Never repeats the token.
      security: [{AdminSession: []}]
      parameters:
        - {$ref: '#/components/parameters/TenantID'}
        - {$ref: '#/components/parameters/RecipientID'}
      responses:
        '200':
          description: Token state.
          content: {application/json: {schema: {$ref: '#/components/schemas/AgentTokenStatus'}}}
    post:
      tags: [Admin configuration]
      operationId: issueRecipientAgentToken
      description: Issues a new assistant token and revokes the previous one. The token is in this response only. The recipient must already have the assistant switch on.
      security: [{AdminSession: []}]
      parameters:
        - {$ref: '#/components/parameters/TenantID'}
        - {$ref: '#/components/parameters/RecipientID'}
        - {$ref: '#/components/parameters/CSRFToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [namesVisible]
              properties:
                namesVisible: {type: boolean, description: When false the assistant sees customer and supplier names as stable aliases.}
      responses:
        '201':
          description: The new token, shown once.
          content: {application/json: {schema: {$ref: '#/components/schemas/AgentIssuedToken'}}}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {description: The assistant is switched off for the recipient or for the whole system., content: {application/json: {schema: {$ref: '#/components/schemas/ProblemEnvelope'}}}}
        '422': {$ref: '#/components/responses/ValidationFailed'}
    delete:
      tags: [Admin configuration]
      operationId: revokeRecipientAgentToken
      description: Revokes the recipient's live token. Revoking when there is none changes nothing.
      security: [{AdminSession: []}]
      parameters:
        - {$ref: '#/components/parameters/TenantID'}
        - {$ref: '#/components/parameters/RecipientID'}
        - {$ref: '#/components/parameters/CSRFToken'}
      responses:
        '200':
          description: The token state after revoking.
          content: {application/json: {schema: {$ref: '#/components/schemas/AgentTokenStatus'}}}
  /api/v1/agent/context:
    get:
      tags: [Agent]
      operationId: getAgentContext
      description: Shop name, today in the shop's time zone, the reports this token may read and the remaining call budget. Read-only; internal network only.
      security: [{AgentToken: []}]
      responses:
        '200': {description: Context., content: {application/json: {schema: {$ref: '#/components/schemas/AgentContext'}}}}
        '401': {$ref: '#/components/responses/AgentRefusal'}
        '429': {$ref: '#/components/responses/AgentRefusal'}
  /api/v1/agent/reports/{reportKey}:
    get:
      tags: [Agent]
      operationId: getAgentReport
      description: The stored numbers of one report for a period, with the period used and when SML was read. A report the token may not read answers exactly like one that does not exist.
      security: [{AgentToken: []}]
      parameters:
        - {name: reportKey, in: path, required: true, schema: {type: string}}
        - {name: dateFrom, in: query, schema: {type: string, format: date}}
        - {name: dateTo, in: query, schema: {type: string, format: date}}
      responses:
        '200': {description: "Report, PREPARING or UNAVAILABLE.", content: {application/json: {schema: {$ref: '#/components/schemas/AgentReport'}}}}
        '401': {$ref: '#/components/responses/AgentRefusal'}
        '404': {$ref: '#/components/responses/AgentRefusal'}
        '422': {$ref: '#/components/responses/AgentRefusal'}
        '429': {$ref: '#/components/responses/AgentRefusal'}
  /api/v1/agent/compare:
    get:
      tags: [Agent]
      operationId: compareAgentReport
      description: The difference between two periods of one metric, computed on the server, with warnings when the periods differ in length, are unfinished or overlap.
      security: [{AgentToken: []}]
      parameters:
        - {name: reportKey, in: query, required: true, schema: {type: string}}
        - {name: metric, in: query, required: true, schema: {type: string}}
        - {name: aFrom, in: query, required: true, schema: {type: string, format: date}}
        - {name: aTo, in: query, required: true, schema: {type: string, format: date}}
        - {name: bFrom, in: query, required: true, schema: {type: string, format: date}}
        - {name: bTo, in: query, required: true, schema: {type: string, format: date}}
      responses:
        '200': {description: Comparison., content: {application/json: {schema: {$ref: '#/components/schemas/AgentCompare'}}}}
        '401': {$ref: '#/components/responses/AgentRefusal'}
        '404': {$ref: '#/components/responses/AgentRefusal'}
        '422': {$ref: '#/components/responses/AgentRefusal'}
        '429': {$ref: '#/components/responses/AgentRefusal'}
...
  /api/v1/agent/alerts:
    get:
      tags: [Agent]
      operationId: getAgentAlerts
      description: The alerts this token's recipient can set, with the threshold and on/off state of each. Alerts are checked once a day by the worker and sent as a message in the assistant's chat; none is sent unless the operator has switched sending on.
      security: [{AgentToken: []}]
      responses:
        '200': {description: The catalog with this recipient's settings., content: {application/json: {schema: {$ref: '#/components/schemas/AgentAlerts'}}}}
        '401': {$ref: '#/components/responses/AgentRefusal'}
        '404': {$ref: '#/components/responses/AgentRefusal'}
        '429': {$ref: '#/components/responses/AgentRefusal'}
  /api/v1/agent/alerts/{ruleKey}:
    put:
      tags: [Agent]
      operationId: putAgentAlert
      description: Sets, changes or switches off one alert for this token's recipient. This is the only write in the Agent API and it changes nothing but the recipient's own alert settings; it is audited. A rule on a report the recipient may not read answers exactly like an unknown rule.
      security: [{AgentToken: []}]
      parameters:
        - {name: ruleKey, in: path, required: true, schema: {type: string, enum: [ar_overdue, ar_over_year, stock_reorder, sales_drop, receipts_drop, margin_drop, morning_digest]}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AgentAlertSet'}
      responses:
        '200': {description: The alert as saved., content: {application/json: {schema: {$ref: '#/components/schemas/AgentAlert'}}}}
        '401': {$ref: '#/components/responses/AgentRefusal'}
        '404': {$ref: '#/components/responses/AgentRefusal'}
        '422': {$ref: '#/components/responses/AgentRefusal'}
        '429': {$ref: '#/components/responses/AgentRefusal'}
  /api/v1/agent/drafts/collection:
    post:
      tags: [Agent]
      operationId: postAgentCollectionDraft
      description: Writes a payment reminder for one of the ten customers owing the most past their due date (no more than a year past due; older debts are reported apart for review), from the receivable report's own figures, for the owner to copy and send. Nothing is sent and nothing is written to SML; the call log records that a draft was made, never its text or the customer. Needs a token that may see names and the receivable report; without the report it answers exactly like a missing report.
      security: [{AgentToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AgentDraftRequest'}
      responses:
        '200': {description: "A draft, or why there is none (PREPARING, UNAVAILABLE, NOT_FOUND, AMBIGUOUS, NOT_CHASEABLE for a debt more than a year past due).", content: {application/json: {schema: {$ref: '#/components/schemas/AgentDraft'}}}}
        '401': {$ref: '#/components/responses/AgentRefusal'}
        '404': {$ref: '#/components/responses/AgentRefusal'}
        '422': {$ref: '#/components/responses/AgentRefusal'}
        '429': {$ref: '#/components/responses/AgentRefusal'}
  /api/v1/agent/lookup/{kind}:
    get:
      tags: [Agent]
      operationId: getAgentLookup
      description: Asks the shop's system one narrow question about one customer or item the master data copy already knows (customer_balance, customer_recent_sales, item_stock), read live and kept for five minutes. The statements are fixed text; the code is the only value used and must be a known record. Limited to two at once and thirty a shop per hour. Customer lookups need a token that may see names. A kind the recipient has no report for answers exactly like a missing report. The call log records the kind only.
      security: [{AgentToken: []}]
      parameters:
        - {name: kind, in: path, required: true, schema: {type: string, enum: [customer_balance, customer_recent_sales, item_stock]}}
        - {name: code, in: query, required: true, schema: {type: string, minLength: 1, maxLength: 80}, description: "A code returned by the search route."}
      responses:
        '200': {description: "The figures, or why there are none (NOT_FOUND, BUSY, UNAVAILABLE).", content: {application/json: {schema: {$ref: '#/components/schemas/AgentLookup'}}}}
        '401': {$ref: '#/components/responses/AgentRefusal'}
        '404': {$ref: '#/components/responses/AgentRefusal'}
        '422': {$ref: '#/components/responses/AgentRefusal'}
        '429': {$ref: '#/components/responses/AgentRefusal'}
  /api/v1/agent/search:
    get:
      tags: [Agent]
      operationId: getAgentSearch
      description: Finds customers, suppliers or items by words of a name, a code or a phone number, in a read-only copy of the shop's master data that the worker makes from SML once a day. Customers and suppliers need a token that may see names. A kind the recipient has no report about answers exactly like a missing report. The call log records the kind only, never the words.
      security: [{AgentToken: []}]
      parameters:
        - {name: kind, in: query, required: true, schema: {type: string, enum: [customer, supplier, item]}}
        - {name: q, in: query, required: true, schema: {type: string, minLength: 2, maxLength: 100}, description: "Up to five words; each must appear in the name, the code or the phone number."}
      responses:
        '200': {description: "The matches (at most ten) or why there are none (NOT_SYNCED, UNAVAILABLE).", content: {application/json: {schema: {$ref: '#/components/schemas/AgentSearch'}}}}
        '401': {$ref: '#/components/responses/AgentRefusal'}
        '404': {$ref: '#/components/responses/AgentRefusal'}
        '422': {$ref: '#/components/responses/AgentRefusal'}
        '429': {$ref: '#/components/responses/AgentRefusal'}
  /api/v1/agent/drafts/purchase-order:
    post:
      tags: [Agent]
      operationId: postAgentPurchaseDraft
      description: Writes a purchase list for the owner to copy and send, from the items below their reorder point in the reorder report (the most short first, up to twenty). The quantity is the shortage against the reorder point, not a recommended order quantity, and the report has no supplier. Nothing is sent and nothing is written to SML; the call log records that a draft was made. Without the reorder report it answers exactly like a missing report.
      security: [{AgentToken: []}]
      responses:
        '200': {description: "A draft, or why there is none (PREPARING, UNAVAILABLE, NOTHING_TO_ORDER).", content: {application/json: {schema: {$ref: '#/components/schemas/AgentPurchaseDraft'}}}}
        '401': {$ref: '#/components/responses/AgentRefusal'}
        '404': {$ref: '#/components/responses/AgentRefusal'}
        '429': {$ref: '#/components/responses/AgentRefusal'}
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: aibcc_admin_session
    ViewerSession:
      type: apiKey
      in: cookie
      name: aibcc_viewer_session
    AgentToken:
      type: http
      scheme: bearer
      description: A per-recipient assistant token issued by an admin. Served on the internal network only.
  parameters:
    TenantID: {name: tenantId, in: path, required: true, schema: {type: string, format: uuid}}
    RecipientID: {name: recipientId, in: path, required: true, schema: {type: string, format: uuid}}
...
    CSRFToken: {name: X-CSRF-Token, in: header, required: true, schema: {type: string, minLength: 32, maxLength: 256}}
  responses:
    AgentRefusal:
      description: A refusal the assistant can pass on. A report that is forbidden answers exactly like one that does not exist.
      content: {application/json: {schema: {$ref: '#/components/schemas/AgentError'}}}
...
    NotFound: {$ref: '#/components/responses/Problem404'}
...
    ValidationFailed: {$ref: '#/components/responses/Problem422'}
...
    Problem404: {description: Resource not found., content: {application/json: {schema: {$ref: '#/components/schemas/ProblemEnvelope'}}}}
...
  schemas:
...
    AgentTokenStatus:
      type: object
      required: [enabled, info]
      properties:
        enabled: {type: boolean, description: False while the assistant is switched off for the whole system.}
        info: {$ref: '#/components/schemas/AgentTokenInfo'}
    AgentIssuedToken:
      type: object
      required: [token, info]
      properties:
        token: {type: string, description: Shown once. Stored only as a hash.}
        info: {$ref: '#/components/schemas/AgentTokenInfo'}
    AgentAlert:
      type: object
      required: [rule, label, description, unit, available, enabled]
      properties:
        rule: {type: string, enum: [ar_overdue, ar_over_year, stock_reorder, sales_drop, receipts_drop, margin_drop, morning_digest]}
        label: {type: string, description: What the owner calls it.}
        description: {type: string, description: "What is watched, in plain Thai."}
        unit: {type: string, enum: [THB, COUNT, PERCENT, POINTS, SWITCH], description: "SWITCH is a rule that is only on or off and has no threshold."}
        available: {type: boolean, description: False when the recipient may not read the report the rule watches.}
        threshold: {type: string, description: "Set by the owner. Baht, a number of items or a percent, as plain digits."}
        enabled: {type: boolean}
        lastFiredAt: {type: string, format: date-time}
        message: {type: string, description: Plain Thai the assistant can pass on after a change.}
    AgentAlerts:
      type: object
      required: [status, alerts]
      properties:
        status: {type: string, enum: [READY]}
        alerts: {type: array, items: {$ref: '#/components/schemas/AgentAlert'}}
        notes: {type: array, items: {type: string}}
    AgentAlertSet:
      type: object
      additionalProperties: false
      properties:
        threshold: {type: string, description: "Omit to keep the current threshold (switching a rule off or on)."}
        enabled: {type: boolean, description: Defaults to true.}
    AgentDraftRequest:
      type: object
      required: [customer]
      additionalProperties: false
      properties:
        customer: {type: string, minLength: 2, maxLength: 100, description: "Part of the customer's name, as the owner said it."}
        tone: {type: string, enum: [friendly, formal], description: Defaults to friendly.}
    AgentDraft:
      type: object
      required: [status]
      properties:
        status: {type: string, enum: [READY, PREPARING, UNAVAILABLE, NOT_FOUND, AMBIGUOUS, NOT_CHASEABLE]}
        message: {type: string}
        customer: {type: string}
        overdueAmount: {type: string, description: "Baht past the due date, as plain digits."}
        maxDaysPastDue: {type: integer}
        tone: {type: string, enum: [friendly, formal]}
        draft: {type: string, description: "The text to copy, word for word. Present only when status is READY."}
        collectedAt: {type: string, format: date-time}
        freshness: {type: string, enum: [FRESH, STALE]}
        candidates: {type: array, items: {type: string}, description: "Names the owner can choose from (AMBIGUOUS, NOT_FOUND)."}
        notes: {type: array, items: {type: string}}
        retryAfterSeconds: {type: integer}
...
    AgentSearch:
      type: object
      required: [status, totalFound]
      properties:
        status: {type: string, enum: [READY, NOT_SYNCED, UNAVAILABLE]}
        message: {type: string}
        kind: {type: string, enum: [CUSTOMER, SUPPLIER, ITEM]}
        totalFound: {type: integer}
        syncedAt: {type: string, format: date-time, description: "When the copy was made from SML."}
        matches:
          type: array
          items:
            type: object
            required: [code, name]
            properties:
              code: {type: string}
              name: {type: string}
              phone: {type: string}
              unit: {type: string}
              supplierCode: {type: string}
        notes: {type: array, items: {type: string}}
...
    AgentContext:
      type: object
      required: [shop, timezone, today, reports, limits]
      properties:
        shop: {type: string}
        timezone: {type: string}
        today: {type: string, format: date}
        reports:
          type: array
          items:
            type: object
            required: [key, label, category, categoryLabel, periodMode, supportsCompare]
            properties:
              key: {$ref: '#/components/schemas/ReportKey'}
              label: {type: string}
              category: {type: string}
              categoryLabel: {type: string}
              periodMode: {type: string, enum: [DATE_RANGE, AS_OF_DATE, CURRENT_ONLY]}
              supportsCompare: {type: boolean}
        limits:
          type: object
          required: [callsPerHour, callsRemaining]
          properties:
            callsPerHour: {type: integer}
            callsRemaining: {type: integer}
        notes: {type: array, items: {type: string}}
    AgentReport:
      type: object
      required: [reportKey, label, status]
      properties:
        reportKey: {$ref: '#/components/schemas/ReportKey'}
        label: {type: string}
        status: {type: string, enum: [READY, PREPARING, UNAVAILABLE]}
        message: {type: string}
        period: {$ref: '#/components/schemas/AgentPeriod'}
        comparisonPeriod: {$ref: '#/components/schemas/AgentPeriod'}
        collectedAt: {type: string, format: date-time, description: "When SML was read, in the shop's time zone."}
        freshness: {type: string, enum: [FRESH, STALE, DELIVERED]}
        deliveredAt: {type: string, format: date-time}
        kpis:
          type: array
          items:
            type: object
            required: [key, label, unit, value]
            properties:
              key: {type: string}
              label: {type: string}
              unit: {type: string, enum: [THB, COUNT, PERCENT, QUANTITY, RATIO]}
              value: {type: string}
              comparison:
                type: object
                properties:
                  previousValue: {type: string}
                  delta: {type: string}
                  percent: {type: string}
                  direction: {type: string}
        visualizations:
          type: array
          items:
            type: object
            properties:
              key: {type: string}
              title: {type: string}
              intent: {type: string}
              unit: {type: string}
              categories: {type: array, items: {type: string}}
              series:
                type: array
                items:
                  type: object
                  properties:
                    key: {type: string}
                    label: {type: string}
                    values: {type: array, items: {type: string}}
        warnings: {type: array, items: {type: string}}
        retryAfterSeconds: {type: integer}
...
    AgentCompare:
      type: object
      required: [reportKey, label, status]
      properties:
        reportKey: {$ref: '#/components/schemas/ReportKey'}
        label: {type: string}
        status: {type: string, enum: [READY, PREPARING, UNAVAILABLE]}
        message: {type: string}
        metric: {type: string}
        metricLabel: {type: string}
        unit: {type: string}
        a: {$ref: '#/components/schemas/AgentCompareSide'}
        b: {$ref: '#/components/schemas/AgentCompareSide'}
        delta: {type: string}
        percent: {type: string, description: Absent when the second period is zero.}
        warnings: {type: array, items: {type: string}}
        collectedAt: {type: string, format: date-time}
        retryAfterSeconds: {type: integer}
...
    ProblemEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId, retryable]
          properties:
            code: {type: string}
            message: {type: string}
            requestId: {type: string}
            retryable: {type: boolean}
            fieldErrors: {type: array, items: {$ref: '#/components/schemas/FieldError'}}
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field: {type: string}
        code: {type: string}
        message: {type: string}
...